Fix tooltips
