+++ app-desktop/src/jvmMain/kotlin/io/stardewvalleydesigner/main.kt
@@ -34,7 +34,7 @@ import io.stardewvalleydesigner.screens.MainMenuScreen
 import io.stardewvalleydesigner.settings.Lang
 import io.stardewvalleydesigner.ui.component.editor.EditorScreen
 import io.stardewvalleydesigner.ui.component.editor.res.WithImageResources
-import io.stardewvalleydesigner.ui.component.editor.screen.bottommenu.*
+import io.stardewvalleydesigner.ui.component.editor.screen.topmenu.*
 import io.stardewvalleydesigner.ui.component.settings.GlobalSettings
 import io.stardewvalleydesigner.ui.component.settings.WithSettings
 import io.stardewvalleydesigner.ui.component.themes.AppTheme
