+++ common/ui-components/editor-menus/save-design-as/src/commonMain/kotlin/io/stardewvalleydesigner/ui/component/editor/screen/topmenu/savedesignas/JsonUtils.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.stardewvalleydesigner.ui.component.editor.screen.bottommenu.savedesignas
+package io.stardewvalleydesigner.ui.component.editor.screen.topmenu.savedesignas
 
 
 internal const val JSON_FORMAT: String = "json"
